pyproject.toml
[project]
name = "DnD_RAG_Chat"
version = "0.0.1"
description = "RAG tool to help do things related to DMing or player character sheet design / homebrewing etc."
authors = [
    {name = "KotoroShinoto", email = "goochmi@gmail.com"},
]
dependencies = ["click>=8.1.8", "urlpath>=1.2.0", "pydantic>=2.9.2", "flask>=3.1.0", "requests>=2.32.3", "unstructured[pdf]>=0.11.8", "langchain>=0.3.14", "langchain-community>=0.3.14", "haystack>=0.42", "llama-index>=0.12.9", "unstructured-client>=0.28.1", "openai>=1.59.3", "libmagic>=1.0", "python-magic-bin>=0.4.14", "PyMuPDF>=1.25.1", "rapidocr-onnxruntime>=1.2.3", "pdfminer-six>=20231228", "haystack-ai>=2.7.0", "flask-session>=0.8.0", "flask-sqlalchemy>=3.1.1", "flask-socketio>=5.5.1"]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "The Unlicense"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
packages = [{ include = "your_package", from = "src" }]
